Rework arcosphere model, for slimmer recipes.
* Changes:

- Switch from concrete types to traits for Set and Recipe.
- Rename RecipeSet to ArcosphereFamily, as a collection of related
  types.
- Adjust all.

* Motivation:

A given family only has a fixed set of recipes. Using an enum is thus
natural, ensuring that validation that a given (parsed) recipe is part
of the family must occur, and as a bonus it's much more lightweight (1
byte, vs 16 bytes for the full representation).

This modelization also allows bypassing a lot of const generics, which
reduces the number of const generics bounds and allays issues with the
fairly spotty support in the compiler for now (it _is_ an experimental
feature, after all).