Introduce StagedPath, exposing parallelism in recipe pipelines.
* Changes:

- Introduce StagedPath.
- Return list of StagedPaths from solver.
- Verify StagedPath in verifier.

* Motivation:

Expose parallelism in solutions.

Parallelism, exposed through stages, is key to the latency of the
pipeline, which directly influences the number of arcospheres necessary
to run a pipeline at a certain pace.

Further, exposing the parallelism helps making it more obvious how to
build the path.

* Limitation:

The verify command does not yet verify the parallelism exposed.

A Stage type (slice of recipes) would make this easier, and will come
later.