Fix selection of shortest path
* Changes:

- Select by minimum number of stages, before selecting by minimum number
  of recipes.
- Adjust tests.

* Motivation:

The number of stages is the critical quantity, timing-size, as it
determines the latency at which the arcospheres are recovered.

The number of recipes still matter, though, as each recipe requires a
Gravimetrics facility to execute it.

* Limitation:

It may be desirable to introduce configuration tweaks, here, allowing
the user to opt for "within a distance of minimum" rather than strictly
equal to minimum.